+++ extensions/shared/library/src/main/java/app/revanced/extension/shared/spoof/SpoofVideoStreamsPatch.java
@@ -1,6 +1,7 @@
 package app.revanced.extension.shared.spoof;
 
 import android.net.Uri;
+import android.text.TextUtils;
 
 import androidx.annotation.Nullable;
 
@@ -10,7 +11,6 @@ import java.util.Map;
 import app.revanced.extension.shared.Logger;
 import app.revanced.extension.shared.Utils;
 import app.revanced.extension.shared.settings.BaseSettings;
-import app.revanced.extension.shared.settings.EnumSetting;
 import app.revanced.extension.shared.settings.Setting;
 import app.revanced.extension.shared.spoof.requests.StreamingDataRequest;
 
@@ -19,7 +19,7 @@ public class SpoofVideoStreamsPatch {
     private static final boolean SPOOF_STREAMING_DATA = BaseSettings.SPOOF_VIDEO_STREAMS.get();
 
     private static final boolean FIX_HLS_CURRENT_TIME = SPOOF_STREAMING_DATA
-            && BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get() == ClientType.IOS;
+            && BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get() == ClientType.IOS_UNPLUGGED;
 
     /**
      * Any unreachable ip address.  Used to intentionally fail requests.
@@ -33,20 +33,6 @@ public class SpoofVideoStreamsPatch {
     private static boolean isPatchIncluded() {
         return false; // Modified during patching.
     }
-
-    public static final class NotSpoofingAndroidVrAvailability implements Setting.Availability {
-        @Override
-        public boolean isAvailable() {
-            if (SpoofVideoStreamsPatch.isPatchIncluded()) {
-                EnumSetting<ClientType> clientType = BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE;
-                return clientType.isAvailable() && clientType.get() != ClientType.ANDROID_VR;
-            }
-
-            return true;
-        }
-    }
-
-
 
     /**
      * Injection point.
@@ -102,6 +88,17 @@ public class SpoofVideoStreamsPatch {
      */
     public static boolean isSpoofingEnabled() {
         return SPOOF_STREAMING_DATA;
+    }
+
+    /**
+     * Injection point.
+     * Only invoked when playing a livestream on an iOS client.
+     */
+    public static boolean fixHLSCurrentTime(boolean original) {
+        if (!SPOOF_STREAMING_DATA) {
+            return original;
+        }
+        return false;
     }
 
     /**
@@ -193,22 +190,47 @@ public class SpoofVideoStreamsPatch {
 
     /**
      * Injection point.
-     *
-     * Fixes iOS livestreams starting from the beginning.
      */
-    public static boolean fixHLSCurrentTime(boolean original) {
-        if (FIX_HLS_CURRENT_TIME) {
-            return false;
+    public static String appendSpoofedClient(String videoFormat) {
+        try {
+            if (SPOOF_STREAMING_DATA && BaseSettings.SPOOF_STREAMING_DATA_STATS_FOR_NERDS.get()
+                    && !TextUtils.isEmpty(videoFormat)) {
+                // Force LTR layout, to match the same LTR video time/length layout YouTube uses for all languages.
+                return "\u202D" + videoFormat + "\u2009(" // u202D = left to right override
+                        + StreamingDataRequest.getLastSpoofedClientName() + ")";
+            }
+        } catch (Exception ex) {
+            Logger.printException(() -> "appendSpoofedClient failure", ex);
         }
 
-        return original;
+        return videoFormat;
+    }
+
+    public static final class NotSpoofingAndroidAvailability implements Setting.Availability {
+        @Override
+        public boolean isAvailable() {
+            if (SpoofVideoStreamsPatch.isPatchIncluded()) {
+                return !BaseSettings.SPOOF_VIDEO_STREAMS.get()
+                        || BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get() == ClientType.IOS_UNPLUGGED;
+            }
+
+            return true;
+        }
+    }
+
+    public static final class AudioStreamLanguageOverrideAvailability implements Setting.Availability {
+        @Override
+        public boolean isAvailable() {
+            return !BaseSettings.SPOOF_VIDEO_STREAMS.get()
+                    || BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get() == ClientType.ANDROID_VR_NO_AUTH;
+        }
     }
 
     public static final class SpoofiOSAvailability implements Setting.Availability {
         @Override
         public boolean isAvailable() {
             return BaseSettings.SPOOF_VIDEO_STREAMS.get()
-                    && BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get() == ClientType.IOS;
+                    && BaseSettings.SPOOF_VIDEO_STREAMS_CLIENT_TYPE.get() == ClientType.IOS_UNPLUGGED;
         }
     }
 }